backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt pyproject.toml
a2wsgi==1.10.8
    # via bentoml
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via bentoml
aiosignal==1.3.2
    # via aiohttp
aiosqlite==0.21.0
    # via bentoml
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   httpx-ws
    #   starlette
    #   watchfiles
appdirs==1.4.4
    # via fs
asgiref==3.8.1
    # via opentelemetry-instrumentation-asgi
attrs==25.3.0
    # via
    #   aiohttp
    #   bentoml
    #   cattrs
bentoml==1.4.6
    # via songsnap-backend (pyproject.toml)
cachetools==5.5.2
    # via google-auth
cattrs==23.1.2
    # via bentoml
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   bentoml
    #   click-option-group
    #   uvicorn
click-option-group==0.5.7
    # via bentoml
cloudpickle==3.1.1
    # via bentoml
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
docstring-parser==0.16
    # via google-cloud-aiplatform
fastapi==0.115.12
    # via songsnap-backend (pyproject.toml)
ffmpeg-python==0.2.0
    # via songsnap-backend (pyproject.toml)
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fs==2.4.16
    # via bentoml
future==1.0.0
    # via ffmpeg-python
google-api-core==2.24.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform==1.71.1
    # via
    #   songsnap-backend (pyproject.toml)
    #   vertexai
google-cloud-bigquery==3.30.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via google-cloud-aiplatform
google-crc32c==1.7.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.7.0
    # via songsnap-backend (pyproject.toml)
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.69.2
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-resource-manager
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.7
    # via
    #   httpx
    #   httpx-ws
httpx==0.28.1
    # via
    #   bentoml
    #   google-genai
    #   httpx-ws
httpx-ws==0.7.1
    # via bentoml
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via opentelemetry-api
jinja2==3.1.6
    # via bentoml
kantoku==0.18.3
    # via bentoml
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.4
    # via
    #   songsnap-backend (pyproject.toml)
    #   bentoml
    #   shapely
nvidia-ml-py==12.570.86
    # via bentoml
opentelemetry-api==1.31.1
    # via
    #   bentoml
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.52b1
    # via
    #   bentoml
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-aiohttp-client==0.52b1
    # via bentoml
opentelemetry-instrumentation-asgi==0.52b1
    # via bentoml
opentelemetry-sdk==1.31.1
    # via bentoml
opentelemetry-semantic-conventions==0.52b1
    # via
    #   bentoml
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.52b1
    # via
    #   bentoml
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
packaging==24.2
    # via
    #   bentoml
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   opentelemetry-instrumentation
    #   pip-requirements-parser
pathspec==0.12.1
    # via bentoml
pillow==11.1.0
    # via songsnap-backend (pyproject.toml)
pip-requirements-parser==32.0.1
    # via bentoml
prometheus-client==0.21.1
    # via bentoml
prompt-toolkit==3.0.50
    # via questionary
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==5.29.4
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psutil==7.0.0
    # via
    #   bentoml
    #   kantoku
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via
    #   songsnap-backend (pyproject.toml)
    #   bentoml
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pyparsing==3.2.2
    # via pip-requirements-parser
python-dateutil==2.9.0.post0
    # via
    #   bentoml
    #   google-cloud-bigquery
python-dotenv==1.0.1
    # via bentoml
python-json-logger==3.3.0
    # via bentoml
python-multipart==0.0.20
    # via bentoml
pyyaml==6.0.2
    # via bentoml
pyzmq==26.3.0
    # via kantoku
questionary==2.1.0
    # via bentoml
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
rich==13.9.4
    # via bentoml
rsa==4.9
    # via google-auth
schema==0.7.7
    # via bentoml
setuptools==78.0.2
    # via fs
shapely==2.0.7
    # via google-cloud-aiplatform
simple-di==0.1.5
    # via bentoml
six==1.17.0
    # via
    #   fs
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via
    #   bentoml
    #   fastapi
tomli-w==1.2.0
    # via bentoml
tornado==6.4.2
    # via kantoku
typing-extensions==4.12.2
    # via
    #   songsnap-backend (pyproject.toml)
    #   aiosqlite
    #   anyio
    #   fastapi
    #   google-genai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via bentoml
vertexai==1.71.1
    # via songsnap-backend (pyproject.toml)
watchfiles==1.0.4
    # via bentoml
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via google-genai
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
wsproto==1.2.0
    # via httpx-ws
yarl==1.18.3
    # via aiohttp
yt-dlp==2025.3.21
    # via songsnap-backend (pyproject.toml)
zipp==3.21.0
    # via importlib-metadata
